root/requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
django-celery==3.3.1
django-celery-beat==2.4.0
django-silk==5.0.2
django-timezone-field==5.0
djangorestframework==3.14.0
gprof2dot==2022.7.29
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
